+++ commands/server/main.go
@@ -7,35 +7,36 @@ package main
 
 import (
 	"context"
+	"flag"
 	"fmt"
 	"log"
 	"net/http"
 	"os"
 	"time"
 
-	"github.com/gorilla/handlers"
 	metrics "github.com/kevinburke/go-simple-metrics"
+	"github.com/kevinburke/handlers"
 	"github.com/kevinburke/rickover/config"
 	"github.com/kevinburke/rickover/models/db"
 	"github.com/kevinburke/rickover/server"
 	"github.com/kevinburke/rickover/setup"
 )
 
-func configure() (http.Handler, error) {
+func configure(ctx context.Context) (http.Handler, error) {
 	dbConns, err := config.GetInt("PG_SERVER_POOL_SIZE")
 	if err != nil {
 		log.Printf("Error getting database pool size: %s. Defaulting to 10", err)
 		dbConns = 10
 	}
 
-	if err = setup.DB(context.Background(), db.DefaultConnection, dbConns); err != nil {
+	if err = setup.DB(ctx, db.DefaultConnection, dbConns); err != nil {
 		return nil, err
 	}
 
 	metrics.Namespace = "rickover.server"
 	metrics.Start("web", os.Getenv("LIBRATO_EMAIL_ACCOUNT"))
 
-	go setup.MeasureActiveQueries(5 * time.Second)
+	go setup.MeasureActiveQueries(ctx, 5*time.Second)
 
 	// If you run this in production, change this user.
 	server.AddUser("test", "hymanrickover")
@@ -43,7 +44,10 @@ func configure() (http.Handler, error) {
 }
 
 func main() {
-	s, err := configure()
+	flag.Parse()
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	s, err := configure(ctx)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -53,5 +57,5 @@ func main() {
 		port = "9090"
 	}
 	log.Printf("Listening on port %s\n", port)
-	log.Fatal(http.ListenAndServe(fmt.Sprintf(":%s", port), handlers.LoggingHandler(os.Stdout, s)))
+	log.Fatal(http.ListenAndServe(fmt.Sprintf(":%s", port), handlers.Log(s)))
 }